Explicit cast to size_t HT Dair Grant

+++ src/metrohash128.cpp
@@ -72,7 +72,7 @@ void MetroHash128::Update(const uint8_t * const buffer, const uint64_t length)
         if (fill > length)
             fill = length;
 
-        memcpy(input.b + (bytes % 32), ptr, fill);
+        memcpy(input.b + (bytes % 32), ptr, static_cast<size_t>(fill));
         ptr   += fill;
         bytes += fill;
